0.1.1 TMP UI
Inicio de blockout de UI de tienda y se agrego TMPpro

+++ Assets/Game/Players/SO IF/WeaponData.cs
@@ -7,7 +7,7 @@ public class WeaponData : ScriptableObject
     public string weaponName;
     public float damage;
     public float fireRate;
-    public float magSize;
+    public int magSize;
     public float reloadTime;
     
     [Header("Projectile Settings")]
